apps/services/recon/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
psycopg[binary]==3.2.3
nats-py==2.11.0
orjson==3.10.7
